+++ package.json
@@ -29,7 +29,7 @@
     "@types/parse-color": "^1.0.3",
     "color-parse": "^2.0.2",
     "dotenv": "^16.4.5",
-    "eslint": "^9.0.0",
+    "eslint": "^8.0.0",
     "i18next": "^23.10.1",
     "mathjs": "^12.4.1",
     "module-alias": "^2.2.3",
@@ -40,7 +40,7 @@
     "ts-dedent": "^2.2.0"
   },
   "devDependencies": {
-    "@lisandra-dev/eslint-config": "^1.1.5",
+    "@lisandra-dev/eslint-config": "^1.1.6",
     "ansi-colors": "^4.1.3",
     "chartjs-node-canvas": "^4.1.6",
     "commander": "^12.0.0",
