+++ src/commands/rolls/dbroll.ts
@@ -1,8 +1,9 @@
 import { cmdLn, lError, ln } from "@localization";
 import { default as i18next } from "@localization/i18next";
 import { EClient } from "@main";
-import {filterChoices, replaceFormulaInDice, reply, rollWithInteraction, title } from "@utils";
-import { getUserFromMessage } from "@utils/db";
+import {filterChoices, reply, title } from "@utils";
+import { getFirstRegisteredChar, getUserFromMessage } from "@utils/db";
+import { rollStatistique } from "@utils/roll";
 import { AutocompleteInteraction, CommandInteraction, CommandInteractionOptionResolver, SlashCommandBuilder } from "discord.js";
 import removeAccents from "remove-accents";
 
@@ -87,65 +88,27 @@ export const rollForUser = {
 		const guildData = client.settings.get(interaction.guild.id);
 		const ul = ln(interaction.locale);
 		if (!guildData) return;
-		let optionChar = options.getString(t("common.character"));
+		let optionChar = options.getString(t("common.character")) ?? undefined;
 		const charName = optionChar ? removeAccents(optionChar.toLowerCase()) : undefined;
 		
 		try {
 			let userStatistique = await getUserFromMessage(client.settings, interaction.user.id,  interaction.guild, interaction, charName);
 			if (!userStatistique && !charName){
 			//find the first character registered
-				const userData = client.settings.get(interaction.guild.id, `user.${interaction.user.id}`);
-				if (!userData) {
-					await reply(interaction,{ content: ul("error.notRegistered"), ephemeral: true });
-					return;
-				}
-				const firstChar = userData[0];
-				optionChar = title(firstChar.charName);
-				userStatistique = await getUserFromMessage(client.settings, interaction.user.id, interaction.guild, interaction, firstChar.charName);
+				const char = await getFirstRegisteredChar(client, interaction, ul);
+				userStatistique = char?.userStatistique;
+				optionChar = char?.optionChar || "";
 			}
 			if (!userStatistique) {
 				await reply(interaction,{ content: ul("error.notRegistered"), ephemeral: true });
 				return;
 			}
+			
 			if (!userStatistique.stats) {
 				await reply(interaction,{ content: ul("error.noStats"), ephemeral: true });
 				return;
 			}
-			//create the string for roll
-			const statistique = options.getString(t("common.statistic"), true).toLowerCase();
-			//model : {dice}{stats only if not comparator formula}{bonus/malus}{formula}{override/comparator}{comments}
-			let comments = options.getString(t("dbRoll.options.comments.name")) ?? "";
-			const override = options.getString(t("dbRoll.options.override.name"));
-			const modificator = options.getNumber(t("dbRoll.options.modificator.name")) ?? 0;
-			const userStat = userStatistique.stats?.[removeAccents(statistique)];
-			const template = userStatistique.template;
-			let dice = template.diceType?.replaceAll("$", userStat.toString());
-			if (!dice) {
-				await reply(interaction,{ content: ul("error.noDice"), ephemeral: true });
-				return;
-			}
-			if (override) {
-				const SIGN_REGEX =/(?<sign>[><=!]+)(?<comparator>(\d+))/;
-				const diceMatch = SIGN_REGEX.exec(dice);
-				const overrideMatch = SIGN_REGEX.exec(override);
-				if (diceMatch && overrideMatch && diceMatch.groups && overrideMatch.groups) {
-					dice = dice.replace(diceMatch[0], overrideMatch[0]);
-				} else if (!diceMatch && overrideMatch) {
-					dice += overrideMatch[0];
-				}
-			}
-			const charNameComments = optionChar ? ` • **@${title(optionChar)}**` : "";
-			comments += ` __[${title(statistique)}]__${charNameComments}`;
-			const modificatorString = modificator > 0 ? `+${modificator}` : modificator < 0 ? `${modificator}` : "";
-			const comparatorMatch = (/(?<sign>[><=!]+)(?<comparator>(\d+))/).exec(dice);
-			let comparator = "";
-			if (comparatorMatch) {
-				//remove from dice
-				dice = dice.replace(comparatorMatch[0], "");
-				comparator = comparatorMatch[0];
-			}
-			const roll = `${replaceFormulaInDice(dice)}${modificatorString}${comparator} ${comments}`;
-			await rollWithInteraction(interaction, roll, interaction.channel, client.settings, template.critical);
+			return await rollStatistique(interaction, client, userStatistique, options, ul, optionChar);
 		}
 		catch (error) {
 			console.error(error);
